Update bottle version to fix getargspec error
On the latest Python version, even with a clean installation, running the script may throw an error because of the outdated bottle package version. Python has depreciated the 'getargspec' and now uses 'getfullargspec' instead. This is fixed in the latter versions of bottle, and updating the bottle version seems to fix the issue without causing any backward compatibility issues with the server scripts.